Workaround in bslmf_typelist for VC2015
The VC2015 compiler supports the features required for
the C++11 branch of the code, but also contains a bug
in the parser for templates that compare two integers
using a < operator, mistaking it for an attempt to
instantiate another template (despite the preceding
identifier not being dependant in any way, and
unmabiguously an unsigned integer).  To complicate
matters, the error reported is on the attempt to close
the surrounding namespace 1500 lines later.  It is not
clear how the parser handles all of the other template
definitions in those 1500 lines while looking for the
closing angle bracket for the template.