Merge Implementation of C++11 movable strings
This commits a manual merge of the .h and .cpp file from the
branch to add C++11 mode support to 'basic_string'.  The test
driver will be merged as a follow-up, as it is a much more
complicated and interleaved merge.  Note that both resolved
merges should be pushed atomically when complete.