First part of move-support testing
This is the first part of theC++11  move-support branch
test driver, merging back towards master.  This deliberately
does NOT include the actual test functions for move tests
for test cases 7 and 9, as they generate a very interleaved
diff between the newly added test functions and the test
functions that would have naturally followed in the .t.cpp
file.  Given the large volume of other cleanup happening
in this branch, we make the first diff as clean as possible,
in a form that will compile but will not link.

The next patch will have the missing test functions, and
the final patch in the sequence will bring in the last of
the cleanups from the reviewed branch.

Note that this has been a manual merge, to ensure that no
work that has been landed on master in parallel is lost.