Fix traits for for native string on Solaris
The Oracle CC compiler has a very odd interaction with
our specialization of the 'bslmf::IsBitwiseMoveable'
trait for 'bsl::basic_string', also specializing the
trait for 'native_std::basic_string' at the same time.
It is not clear why this happens, and happens for only
the CC compiler, although we suspect a compiler bug
where it is not obvious what a minimal reproducible
test case would be.  The simple fix is to use the
'BSLMF_NESTED_TRAIT_DECLARATION' macro to supply
the trait instead.