Fix narrowing conversion warnings from template test facility
This patch fixes an annoying warning from gcc complaining of narrowing
conversion when constructing 'char' test data using the template test
facility.  As this warning arise in the <bslalg_scalarprimitives.h>
header, it is both noisy and misleading.

As a drive-by fix, add missing assertions that the identifiers passed
to 'emplace' overloads are in the range specified by the contract.
This is an important check for the character overloads added by the
patch.

Finally, we choose to document these overloads directly in the class
definition.  The existing precedent set for the member-pointer test
type is to provide an (undocumented) explicit template specialization
for the 'emplace' method below-the-line in the implementation details
of this class.  Doing this kind of specialization advertises how, in
rare circumstances, the author of a test driver could add their own
test type to the set supported for generation by the test facility.
However, despite supplying this example, we do not document the
general scheme for extension, which works only for specific types as
there is no such thing as a partial template specialization for a
function template, and the class interface is otherwise closed for
extension, such as supplying an addtional overload (the usual
alternative to partial specialization).  This is also the reason
that generic pointers ('TYPE **') are overloaded directly in the
class.  We opt for the overload, rather than specializing the
function template, as it is more conventional and the extension
scheme is not actually documented.  This decision could easily be
revisited in a subsequent patch.

Note that the explicit specialization for method pointers is
calling out to be moved into the .cpp file, as bde_verify warns
that it is well over our usual threshold for an inline function,
but as an explicit specialization appearing in the header, it
must be defined with the 'inline' keyword.