Fix warning in ball_userfieldsschama.h
Warning on converting a 'size_t' to an 'int', which could
be a significant narrowing on a 64-bit platform.  That said,
2 billion field values per object  should be more than enough for
much longer than the immediate future, so the easiest fix is
to cast to 'int', rather than change the return type.

Considered adding a 'BSLS_ASSERT_SAFE' that the value was
less than 'numeric_limits<int>::max()', but did not want to
add an extra header dependency in this header.  Would have
done so if the method were out-of-line in the .cpp file.