Fix byte-count data type to match allocators
'int' is no longer a suitable type to track memory usagae, in
a 64-bit world.  Use the 'size_type' from the underlying
allocator, which maps to 'std::size_t' and will have the
appropiate size on all platforms, although also switches from
signed to unsigned.  This resolves a number of warnings
originating in a header file.