Resolve narrowing warning in bslst_stringref header
Strictly speaking, this patch silences a warning that is giving
us useful information on a 64-bit platform, when trafficking in
references to strings longer than 2-billion elements.  In practice,
that is not something we will need to worry about for a very long
time, so we silence the warning in favor of existing practice.