bslstl_vector.t.cpp static_asserts in C++11
The revised logic for 'bslmf::IsBitwiseMoveable' causes a
static_assert to fire in a C++11 build for a type that
would be flagged as bitwise-movable in C++03 (size 1) but
having a non-trivial copy constructor detected in C++11.
This patch deals with an issue with the test types in the .t.cpp.