Detect a 'nullptr' macro breaking our headers
We have received a couple of bug reports against 'bsls_nullptr.h'
that turn out to be cause by some intervening header defining a
'nullptr' macro, possibly in an attempt to look more C++11ish.
The cause and effect are not obvious though, so add a simple
#error check for the troublesome macro in C++11 builds, where
it is both illegal, and causes problems.  The C++03 code path
is not affected.