Disable native type traits support on Windows
There is a problem where the #include for <type_traits>
goes recursive in BSL_OVERRIDES_STD mode, when the MSVC
implementation includes <cstddef>, and the recursive
epilog of our own incercept header triggers recursion
in the middle of include-ing <bslmf_integralconstant.h>

As a short term patch to re-enable testing on Windows,
we disable native traits support.  A subsequent patch,
once the recursion has been eliminated, will restore
the facility.