Remove unused header from bslstl_sharedptr.h
Remove allocator_traits from this header, as it is not used directly
by the shared_ptr component, but rather is separately used by the
different 'rep' objects when they construct shared objects.

As the next release of BDE is looking to migrate allocator traits
down to bslma anyway, and clients transitive use will still be
satisfied by the transitive includes of the rep objects, we should
be safe to make this clean removal here and now.