Disable a test on sun studio due to a bug
Disable a test that fails on Sun to due a bug bug in Studio Studio's C++
standard library: 'ostreambuf_iterator' doesn't set the 'failed'
attribute when the iterator reaches the end of EOF of the 'streambuf'.
Therefore, 'operator<<' for the decimal types does not set the 'fail'
and 'bad' bits when streaming to an 'ostream' with a 'streambuf' that is
not large enough.  This is consistent with the behavior for 'int' and
'double'.  Note that the bug does not exist when using stlport.