Fixing missing pcre depends
As a user of libbdl-dev, I want to use the generate pkg-config.
The generated pkg-config file for bdl currently references pcre
on the link line; however that is not a depends on the bdl library
like inteldfp (for example) and thus fails.

In practice, this means that pcre library is not in the robo exposed
refroot.