Fix check for (core) undefined behavior in deque iterators
This constructors performs a range-check on the supplied pointer to
confirm that it lies within the expected data block, but the test
that it must be at the base address or higher is trivially true, or
undefined behavior.  This is triggering a gcc warning in opt builds,
which must now be resolved to continue building with -Werror.

The solution is to cast to integers, and compare the integer values
instead.  This is clearly a pessimization as we pay for a check that
would otherwise optimize away - on the other hand, that is the whole
point of this SAFE check.