Switch new traits inclusion to be opt-in rather than opt out.
We have too many confusing platforms that don't have standard libraries that
support the full panoply of type traits, and also do not have indicator macros
to tell us what they're missing.  So I changed the header to require opt-in
rather than opt-out; it assumes that the is_trivially_... traits are missing
except for some specific combinations of compiler and platform.

The indicator macros that are set for this are not changed if they're already
defined and the tests that use them check whether they're zero, not whether
they're defined, so it's possible to set them on the command line with -D.