Support BSLS_BSLTESTUTIL_ASSRT in if/else clauses
Wrap the ASSERT macros with a redundant 'do {} while(false)' to require a terminating ';'
be supplied by the user.  This supports 'if (a) ASSERTV(blah); else ASSERTV(blag);' as
acceptable style, rather than forcing the ASSERT branches be wrapped in {braces} by the
user.  If does require all ASSERT expressions be terminated by a ';' which may break
existing test drivers - indeed, several 'bsl' test drivers needed fixing to support this
patch.

As the use of 'bsls_bsltestutil' should the limited to the 'bsl' package group itself,
this /should/ be a safe patch to roll out.  Considerable thought should be given before
appliying a similar patch to 'bslim_testuitl'.