Fix hash test driver for VC2015
The problem turned out to be 'is_integral' not being updated to support
C++11 unicode character types, after the test driver for hashing was
updated to support them.  This highlighted a real bug, that the hash
support relied on a correct result from 'is_integral'.

As work on 'is_integral' was necesary, provided clean-up for some
testing issues, including other unaddressed concerns, and specifically
implenment/test in terms of 'long long' and 'unsigned long long',
rather than 'bsls::Types' typedefs, that are not strictly required to
provide a full set of the fundamental types in the language, which IS
the purpose of this component.

Finally, propagate an update back into 'bsls::Types' that 'int64' on
Windows, for all of the compiler versions that we support, is indeed
the C++11 'long long' type, just as it is on all of the Unix C++03
compilers that provide this extension throguh their C99 support.