Support null pointer literals in bslma allocator
The 'deleteObject' factory protocol supports null pointers within the basic
contract, when called with a pointer that happens to be null.  However, it
fails to handle null pointer literals, as they do not deduce to a single
pointer type.  This turns out to be a problem as we start trying to support
more C++11 idioms, in particular, the SFINAE-based constructors in 'shared_ptr'
and elsewhere that the standard library specifies.

The 'deleteObject' factory protocol is a BDE extension, and as such, we are
free to choose to support null only for true pointer types.  However, as null
values are supoprted, it seems an odd inconsistency, and on balance, support
for null pointers in general seems the best way forward.

A respectable alternative would have been to narrow the contract on the
original 'deleteObject'.  However, we have no easy way to know how much
deployed code is relying on that part of the contract, and retroactively
narrowing a deployed contract is the worst kind of way to introduce subtle
bugs into previously working code.   Hence, the decision to support
'bsls::nullptr_t' overloads instead.