Update is_copyconstructible for all MSVC toolchains
The is_copy_constructible trait was getting wrong answers on
some MSVC tool-chains, notably those that provided an implementation
of the native trait, but one that gives the wrong answers.

Along the way, updating the tests stumbled into DRQS 81567505, which
is also addressed by this patch.  This, in turn, required a quick
update to 'bsls_compilerfeatures', which will also be relied upon
by subsequent patches for testing the updated 'shared_ptr' branch.