Fix traits with broken tests on Windows
Several test drivers for type traits test references to array types, where
the array has an unknown bound, e.g., 'int (&)[][42]'.  The VC2010 compiler
does not support such references, and fails to build (and run) the test
driver.  It is reasonable to not test such types on this compiler, as the
user cannot make use of them either.

In preparing this patch, the usual clean-up for bde_verify was added.
Two of these traits were recent additions from Pablo, and so had not been
subject to a previous round of bde_verification, resulting in a little
more noise in the change-set than might be desired.