Fix tests of incomplete key types in unordered maps
The test drivers for unordered (multi) maps fail to build in SAFE_2
mode due to an invariant check in an implementation-detail component.
This is actually a bug in the test driver, in that there is a basic
requirement that the key-type be a complete type, and be hashable -
at least in order to satisfy this sanity check.  In practice, the
test driver gets away with this in non-SAFE_2 builds as long as no
elements are inserted into the test case for incomplete types, so
we restrict the fix to only SAFE_2 builds.