Fix declaration of bool constant types for MSVC2010
When writing test drivers to validate perfect forwarding, the preferred idiom
involves declaring some constant objects of either 'true_type' or 'false_type'
according to some compile-time parameter determined by instantiating the test
case template.  In C++11 these objects are best brace-initialized, in order
to avoid accidentally declaring (and not defining) a function under the most
vexing parse interpretation.  Unfortunately, 'integral_constant' is not an
aggregate in BDE, so the intitialization is not well-formed for a C++03
compiler, such as MSVC2010.  Conversely, if we are going to use this idiom
widely across test drivers, there is an even cleaner and simpler form that
will become available in C++17.

This patch replaces the object declarations and definitions with a macro
that will expand to the appropriate declaration/definition for the language
features available on the platform being compiled.