Revised for review feedback
Rather than fix the broken extenral header-guard, a different solution was
applied for the original problem.  The dependency on 'std::swap' is entirely
removed, which avoid the problem of which standard header it lives in varying
according to which standard is being implemented.  This allows us to move the
<algorithm> dependency down into the transitive includes block.

Running 'bde_verify' to catch any other inadvertant issues highlighted a few
quick fixes (applied) and a lot of churn for alphabetical ordering, names of
parameters being properly used in contracts, and so forth.  Such comments
were deferred to another day.