Address review feedback, plus test driver cleanup
This adds additional tests that 'bsls::Types' typedefs are aliases
for integral types, even though we have no guarantee which such types
are aliased.  This retains a guarantee implied by the previous
implementation, even if it is an unusual place to be asserting the
properties of a lower level component.  This order of testing is
actually mandated by the levelization of the facilities capable of
implementing said tests.