Fix for 32-bit AIX atomics
Due to an odd quirk of architecture, the implementation of theAIX 32-bit atomics
component does NOT #include the compoent header from the .cpp file.  With the
other patches on this branch, this missed that that specific .cpp file had an
undeclared (transitively satisfied) dependency on <bsls_platform.h>.  This is
not expected to raise further concerns.

In the review, found one other header with an undeclared dependency on
bsls_platform, which is transitively satisfied, but now explicitly included to
avoid future concerns.