Add missing test for rvalue reference support
While testing for rvalue references with native traits, forgot to
add the check confirming that rvalue-references are enabled on that
platform.  Due to certain issues with the implementation, we flag
VC2010 and VC2012 as NOT supporting such types, even though they
have a native traits implementation, and a close approximation of
valid syntax that would otherwise be good enough to satisfy this
test driver.