Fix deeper Windows problems with 'is_fundamental' trait
Fixing up the corners and testing the 'is_enum' trait with MSVC 2012 revealed
some deeper problems in the 'is_fundamental' trait that were documented, but
not adequately tested.  This patch adds those tests, and resolves the outstanding
documented problems - although support for 'char16_t' and 'char32_t', while
tested, will not be implemented until a patch up for review in parallel
provides a fix for the 'is_integral' trait.