Deprecate remaining 'bslalg' traits
The various type trait components in 'bslalg' exist only to support
the deprecated 'bslalg_typetraits' component, but have not previously
been separately deprecated.  This became a more significant issue
when it was noticed that documentation of the 'bslalg' traits was
slowly diverging from the traits that they forward to, and we are
getting support request asking which version of the trait (and its
contract) should be used.

This patch then proceeds to replace all use of deprecated traits,
and traits macros, with the preferred trait and declaration syntax
in each case.