Remove C++11 support from gcc 4.7 and earlier
While gcc was tracking the evolving C++0x standard, it incrementally
added C++11 (to be) features under an experimental flag.  However, as
the standard evolved, the semantics of various features changes, and
it is now counterproductive to try to track fine-grained behaviour
changes in unsupported experimental dialects of C++11, so we should
insist on at least gcc 4.8 before enabling C++11 support.