Remove Sun workaround that breaks with Sun CC12.4
The workaround created for the Sun CC compiler is detected as
invalid C++ with the 12.4 compiler, and must be removed.  Rather
than perform a platform check for the version. we prefer to
provide a simpler, clean implementation that will work on all
compilers without any platform checking.  This implementation
uses friend functions to declare regular free function
operators with parameters that happen to be instantiations of
a template type.  This is possible only by defining the
operator functions inline in the class at the point of
declaration - there is no syntax to define such functions
if only forward declared.

A few minor fixups for bde_verify are also applied at this
time, and the test driver is pruned of lots of noise that
was declared by copy/paste from a different component. but
never defined.