Enhance Pablo's test case to validate return value
The last corner of the MSVC bug to validate is that calling the
wrapped (member) function pointers does not corrupt the stack.
Adding an 'int' data member to the nested POD class and verifying
the expected return value completes the set of concerned to be
tested.

As a drive-by fix, remove the unused 'dbg_print' functions that
were adding noise at the top of the test driver.  A follow-up
patch will add the traditional checklist for the test cases.