(almost) full support for C++11 library detection
This a revised patch of work created by George Laptev that
provides feature-detection macros for subsets of C++11 standard
library support with various compilers in our list of supported
tool-chains.  These macros, intially intended for internal BDE
use, advertise which standard library facilities will be available
in namespace 'bsl'.

Note in particular that each macro comes with a prefix denoting
which standard version is being supported.  This allows for future
extensions where C++14 and C++17 extend the 'tuple' and 'pair'
interfaces, for example.  We also start noting which library features
are /removed/ by later standards.

None of the macros in this header are defined unless the current
build mode has at least experimental support for a C++ standard
later than the 2003 version.

Note that the test driver still needs updating with full test
coverage that the various features are indeed available when the
corresponding macro is set.  That has been deferred on the basis
of manual test coverage through higher level components, with a
wide variety of tool-chains.