Third round of CC 12.4 fixes
This addresses almost entirely C++03 builds, in the default
'sun03' build mode.  With this set of patches and workarouds,
we should get a full build of 'bsl', so that we can see the
test drivers that are now failing at runtime.  We will probably
need compiler fixes for these issues before we could roll out
this compiler to production.