Fix exception specification on operator new
In C++11, the exception specification on 'operator new' changes
from 'throw (std::bad_alloc)' to a generic can-throw-anything
specification, and on 'operator delete', it is rewritten using
'noexcept' (although that should be compatible until C++17).

The Sun CC 12.4 compiler is our first compiler to diagnose
mismatches here, but it is strictly correct, so we supply
a patch that does not depend on the specific compiler, but
only on the detected compiler features.