Sun CC 12.4 has 'contexpr' bugs
While 'constexpr' mostly works at the first attempt, there
are corners that still crash the compiler or otherwise fail
in ways that are exposed by BDE test drivers.  We choose to
mark the feature as not-supported for BDE use until it is
fixed (we think) in the subsequent 12.5 release.

Test case added to highlight the kind of problem that must
be resolved before we re-enable support.