Oracle CC 12.4 occasionally asserts on default-template SFINAE
There is an obscure corner-case with default template arguments
used for SFINAE in C++11 that is unfortunately triggered by a
common use case with 'shared_ptr' on the Oracle CC 12.4 compiler.
We need to disable use of this feature in BDE until a patch is
available.  It is believed that this is already fixed in the
subsequent 12.5 compiler.