Support and testing for non-assignable vector elements
Many (but not all) vector operations are required to support
types with private or deleted assignment operators.  Testing
this support has turned up a number of other test coverage
issues, the simplest of which are resolved in passing.

The 'bslstl_vector.t.cpp' test driver no longer builds (again)
on gcc/Linux due to too many template instantiations violating
some internal file limit for an '.o' file on Linux.  Rather
than scale back test coverage again, we split the test driver
into two, with a second component to run the test case above
the classic value semantic type test cases 1-10.  In practice,
we keep test case 11 in the main component, as rounding out the
allocator concerns for test cases 1-10, and for didactic
purposes, retain the usage example.  Original test numbering is
maintained through the Compoent-level test plans across both
components.  Note that test case 8 used for be for the unused
'g' function, but the 'swap' test case was moved from case 19
back into this canonical position.

The restored test coverage quickly revealed that the insert'
method taking an 'intializer_list' was missing from the
specialization for 'const VALUE_TYPE *', so the additional
coverage immediately proved its worth.  The missing method is
now implemented and tested.

Negative tests all failed in 'bslstl_vector_test.t.cpp' as the
component name no longer matched the name of the component that
is asserting.  This is resolved by making all negative tests
RAW, although that also reduces the rigor of our coverage.

In addition, support for vectors of pointer-to-members was
discovered to fail on IBM not due to a compiler bug, but due
to a bad assumption that trivially default constructible
types have an all-zeroes representation for their null value
on all platforms.  The fix for this turned out to be needed
in 'bslalg_arrayprimitives' and 'bslalg_constructionutil',
and we should follow up to fix the test drivers for all other
containers that falsely fail to support this failure (which
we never documented for consumers of the component).  Test
coverage in this component is enhanced to include
'emplace_back()' without any arguments, to verify the correct
prt-to-member fix is in place.

Finally, investigating some of these issues, a little BDE
style cleanup was applied to some components that were under
investigation - notably removing use of deprecated names, and
ensuring that that there were no transitive dependencies on
lower level headers.