Confirm vector support for function pointers
There has been a long-standing issue that 'bsl::vector' does
not support function pointers.  Updating the test driver
confirmed that this is restricted to problems with the
'emplace' functions, and this patch provides the necessary
fix, by simplifying the generated 'emplace' overloads for
vectors of pointers, and then tidying up the test driver
accordingly.

A final audit that all test cases are consistently using
the variety of pointer types, and we're done!