Fix for template instantion bug on Sun
This fixes an odd issue instantiating a particular function
template in a single test driver on Sun CC 12.3, which is
our production compiler.  We have confirmed that the bug is
not present in 12.4 or 12.5, so we are not going to try to
file a reduced test case with Oracle, and leave the workaround
documented in the hopes of removing it along with support for
CC 12.3 one day in the future.