Remove use of deprecated 'auto_ptr'
Gcc now warns of use of the 'auto_ptr' template that was deprecated in
C++11, and removed in C++17.  The warnings on this usage in 'bdlmt'
actually find a real bug, as the 'auto_ptr' will attempt to delete a
single 'char' that was not necessarily allocated with 'new', rather
than return the whole buffer to the supplied allocator.  Fixed using
appropriate 'bslma' tools.