Deprecate the 'BSLS_EXCEPTION_SPEC' macro
Dynamic exception specifications are deprecated in C++11. and removed
from the language in C++17.  We should not be actively supporting
their use with a BDE macro, as conventional wisdom is that they were
a misfeature, even 15-20 years ago.

This patch uses the recently deployed bsls_deprecate component to mark
this macro as deprecated from BDE 3.3 onwards.  We retain support for
the distinct 'throw ()' macro form, as this remains important for
implementing exception classes derived from 'std::exception' in C++03.
We should prefer expanding this to 'noexcept' when the language
feature is available though.  We cannot defer to the BSLS_CPP11 macro
that does exactly this, as we continue to need to expand to 'throw()'
rather than an empty string when 'noexcept' is not available.

As part of the support for deprecating and removing this support, add
a feature-detection macro for the language facility as it is going to
be removed in a coming standard upgrade.