DRQS 65708127: traits support for abominable functions
This is a corner case that looked simple to resolve for a conforming
C++98 compiler, but naturally proved more complicated to work around
issues with our various non-standard compilers.  In the process, we
uncovered a bug that our existing implementation also treats function
references as function types.

This patch greatly increases the test coverage, to find several false
positives as well as correcting support for abominable function types
in 'is_function' and 'add_pointer'.  There are minor limits on the
supported range of function types on IBM, which still supports only
function of up to 14 parameters (or any number of parameters for an
abominable function type), and for Sun which does not correctly
report as functions the extreme corner case of 'const volatile'
qualified abominable functions that have more than 14 parameters.

As this patch fixes several previously undetected false positives,
it may expose errors in upstream components.  In particular, the
'bslmf_detectnestedtrait' component needed adjusting to correctly
detect (or rather compile) that function references never have a
nested trait.