Disable pointer-to-member function for bsl::function on IBM
There are some really odd stack corruption issues on IBM when
calling a pointer-to-member stored in a 'bsl::function' object,
that arises from the troublesome test case created (and resolved)
for the Microsoft compiler.  The short-term fix is to use a
'BSLMF_ASSERT' to force a compile-time error instead.  This is a
big hammer though, as it disables all pointer-to-member support,
where the majority of such usage would work correctly.  This is
because there is no easy way in the type system to distinguish
between the good cases and the bad - this is entirely a run-time
issue on a specific compiler.