Remove deprecated 'shared_ptr' constructor
The 'shared_ptr' constructor taking a single pointer to a 'Rep' object
was deprecated around BDE 2.22 time, as it conflicts with the C++11
constructor taking a 'std::nullptr'.  Now that we are trying to support
the full C++11 specification, the work-around of using this constructor
as a substitute for the 'nullptr_t' constructor no longer works, as it
has the wrong exception specification and should not be made 'constexpr',
so it is time to remove this legacy, and ticket anyone with a remaining
dependency on this constructor.

For code that still requires this facility, the correct solution is to
call the constructor taking a 'Rep' and a pointer to the (potentially
aliased) pointer to the object that the 'shared_ptr' should refer to.