Extensive test enhancement through 'bsltf'
This patch extends the range of basic types used in the 'bsltf'
test macros to cover 'const char *' and move-enhanced (but not
move-only) types.  This improves test coverage of all of the
'bsl' containers, and exposes several interesting corner cases.

Test support for 'const char *' required additional testing to
ensure that each generated value in the standard array of test
values would point to a null-terminated string, otherwise the
BDE standard test macros 'P', 'P_', and failing 'ASSERTV' would
segfault trying to print an invalid string.

This change-set also incorporates the deprecation of
'bslalg_scalardestructionprimitives' in favor of the identical,
but lower level', 'bslma_destructionutil'.  This levelization
was necessasry as part of the generalized allocator support
moving down to 'bslma' in BDE 3.0.

Finally, after fixing a number of issues exposed by the newly
enhanced test coverage, there are still some corner cases in
'vector' that need to be addressed, mostly related to the
move-enhanced types.  This will need to be addressed at the
'bslalg_arrayprimitives' level, and some arithmetic in the test
driver also needs adjusting to allow for such types moving now,
rather than copying and allocating.