Reduce warnings in bitset test driver
Just a few lines in bslstl_bitset.t.cpp lead to an inordinate
spew of warnings, due to the number of template instantiations
repeating the warnings.

This patch cleans up all but the last warning in this component,
as it was not obvious what a clean fix for the remaining cast
would be without obscuring the usage example, where clarity
might matter more than obscure compiler warnings.