Remove unnecessary 'using namespace std' from test drivers
Apparently we get an implicit declaration of 'namespace std' if a test
driver has a 'using namespace std', but does not (transitively) include
a header that provides that namespace.  Clang is the only compiler to
warn about this situation, and this is a minimal patch to help clean up
builds using Clang.