Fix buffer overrun in StringRefData test driver
Error caught by warning in gcc7.  By assiging a string literal to a
'const char *' variable, 'sizeof(string)' is actually the size of the
pointer object, not the pointed-to character array.  In practice, the
test driver is fine on 32-bit platforms as the string is longer than a
32-bit pointer, but it will overrun on a 64-bit build as the string is
just a little too short.

Fixed test driver to use a char array for that local variable, so that
the correct size is computed, delivering the original intent of the
test driver.