Fix untested test-case for 'bitset'
Likely due to a copy/paste error, test case 13 for bslstl_bitset
was somehow numbered as test case 34, so was not being run by
WAF.  It turns out that this is also test case zero though. so
was being run by the cmake test runner, and that showed a basic
problem with the test case.

This case tests the 'noexcept' operator on platforms where it is
available, but several tests relied on calling 'operator[]' to
create a 'bitset::reference' to test the 'noexcept' behavor of
the 'reference' type.  Unfortunately, 'operator[]' itself is not
marked 'noexcept' (it has a narrow contract) so these tests all
fail.

Fix the test case numbering so that the test is correctly run
now, and fixed the test to verify the necessary properties
correctly.  Some minor 'bde_verify' cleanup on related code
applied.