Test case for negative testing relies on undefined behavior
Negative testing installs an assertion handler that throws a
well-known exception type.  This is fine for most cases, but
does not play well with 'noexcept', and hence destructors.

It turns out there is a second special case highlighted by
the test driver 'bdlma_pool.t.cpp, as overloaded delete
operators also have undefined behavior if exiting via an
exception.  This turns out to be benign undefined behavior,
for the purposes of our test drivers, on most platforms.
However, MSVC started forcing an 'abort' for arbitrary
delete function exiting via an exception in VC 2015.  We
need to make an exception in the test driver for that
platform, and keep our eyes open for similar failures on
other platforms in the future.