Remove old MSVC workaround from test baljsn test driver
Older Microsoft run time libraries had a feature in the 'snprintf'
function that would often pad a leading zero into the exponent of
a floating point number when rendered in scientific notation.

That particular quirk was removed in VC 2015, and the test driver
needs updating to accomodate the toolchain working as expected!

As a drive-by fix, several bde_verify issues were fixed, mostly
with relation to missing transitive includes, and correct header
order.