Remove old MSVC workarounds for balxml test drivers
Older MSVC toolsets would pad the exponent of a floating point number
with a leading zero, prior to MSVC 2015.  This is accounted for with
some #if workarounds in a couple of 'balxml' test drivers, and that
workaround must now be version checked to not apply to more recent
versions of the toolchain.

As a drive-by fix, correct the test banners in one of the test drivers
to consistently print upper case, resolving a number of 'bde_verify'
concerns.