Less strict validation for matching an OS string on Windows
The string constructed to represent the Service Pack release
on Windows will often not match the string directly reported
by Windows itself, as the BDE string will report '1.0' rather
than just '1'.  Modify the test to look for simply the same
leading string, and assume that trailing text does not matter.
This is for display only, and precise values matter less.
Note that this test may be expected to fail if run on a Windows
OS on a non-English speaking locale.

Fix up a number of issues for bde_verify, and general test
driver best practice.