Fix gcc warning in bslmf_forwardutil
The workaround introduced for an MSVC compiler bug, to avoid creating a
temporary object and returning a reference as it immediately dies on the
stack frame, produces a new warning on gcc that falls fould of our
warnings-as-errors policy set up for cscheckin.

This patch restores the original 'const_cast' behavior for all
supported platforms other than Microsoft.  Not that this excludes
Solaris and AIX compilers, as the branch is available only for
compilers support C++11, specifically rvalue references.