DRQS 121977756: Remove GNUC major constraint.
The `BSLS_DEPRECATE` macro is undefined when compiling C++ code in any
mode prior to C++11 (`__cplusplus < 201103L`), but only for versions of
GCC <= 5.  When any newer version of GCC is compiling code by specifying
`-std=gnu++98`, it will error if `BSLS_DEPRECATE` is used because it
will fail the version check and attempt to use the `[[deprecated]]`
attribute in C++98 mode.