Hide gcc warnings on bitwise moveable
gcc8 is finally warning on using 'memcpy' and 'memmove' to blit
non-trivial types around memory.  This may be a precursor to
exploiting undefined behavior by ignoring our calls.  For now,
the warning can be silenced with an explicit cast to 'void *',
but there is no reason a smarter compiler in the future might
not peer through the cast and act on the undefined behavior
again.

This initial patch should fix everywhere the warning is raised
that is not a test driver.  There are still extensive patches
required to make the test drivers compile warning-free as well.