Disable move-only integration test for MSVC 2013
MSVC 2013 retains a compiler bug (fixed in MSVC 2015) that
copy constructors are not implicitly deleted when implicitly
declared for a class with bases or members that have deleted
their copy constructor.  The test is updated for MSVC 2013 to
show the bug in action by asserting that 'is_copy_constructible'
is true for both a 'unique_ptr' instantiation, and a 'pair' with
a 'unique_ptr' member.