Document 'bsls::Util::forward'
It appears that the 'forward' function in 'bsls_util' has not been
documented, relying intstead on everyone's common assumption that
it means whatever the ISO Standard version means.  Add the missing
doc, along with a warning about using rvalues beyond the immediate
expression  (which was an actual failure in the test driver for
quite a while).